Add webhook model and update client tests

+++ models/shipping_line.go
@@ -2,12 +2,12 @@ package models
 
 type ShippingLine struct {
 	ConektaBase
-	ID             string  `json:"id,omitempty"`
-	Amount         float64 `json:"amount,omitempty"`
-	TrackingNumber string  `json:"tracking_number,omitempty"`
-	Carrier        string  `json:"carrier,omitempty"`
-	Method         string  `json:"method,omitempty"`
-	ParentID       string  `json:"parent_id,omitempty"`
+	ID             string `json:"id,omitempty"`
+	Amount         int64  `json:"amount,omitempty"`
+	TrackingNumber string `json:"tracking_number,omitempty"`
+	Carrier        string `json:"carrier,omitempty"`
+	Method         string `json:"method,omitempty"`
+	ParentID       string `json:"parent_id,omitempty"`
 }
 
 type ShippingLines struct {
